Refactor values out of settings
This moves towards allowing a setting to have multiple values (one per configuration), which hopefully will allow us to remove the 'configuration' and 'platform' context options from the UI altogether.

A setting's values may be either configured (potentially many) or unconfigured (singular).

Templates are updated to allow showing mutliple values for a single setting.

This also introduces the as-yet-unfinished concept of SettingCommands, which will be displayed in a menu on settings and allow toggling between configured/unconfigured states for a setting (and potentially other actions in future).

This change regresses the conditional relationship between settings previously achieved by nesting settings within BoolSetting/RadioSetting. This can be addressed in another way.

Introduce FontAwesome for a glyph (prototype only) and immutable collections packages.