Move margin to StackPanel
Previously the hit test could hit the bottom of the group area. It has no idea whether it's hit the top or bottom of the area. When we hit the top, we want to consider it as the first setting in that group. However we were doing the same when hit testing the bottom.

By moving the margin to the stack panel, we maintain the same layout but the hit test does not succeed as the bottom margin is not considered part of the group.