Don't respond to own change event
When the settings list scrolls, it updates the CurrentSection property as needed. However the list also responds to changes in this property's value from other sources (i.e. the navigation tree). This would result in us jumping the scroll position, which is confusing to the user.

This commit suppresses one occurrence of the change handler when we set the value ourselves.